fix: replace asset log trigger to asset log crud service
