+++ exam_django/asset/service/asset_log_crud_service/asset_log_service.py
@@ -4,6 +4,11 @@ from user_auth.models import User
 from response import APIResponse
 from messages import ASSET_NOT_FOUND, ASSET_LOG_FOUND
 from rest_framework import status
+from django.db.models.signals import pre_save
+from django.dispatch import receiver
+from django.db import transaction
+from asset.models import Asset
+
 
 class AssetLogService:
     @staticmethod
@@ -124,3 +129,29 @@ class AssetLogService:
                 message=ASSET_NOT_FOUND,
                 status=status.HTTP_404_NOT_FOUND,
             )
+
+
+@receiver(pre_save, sender=Asset)
+def log_asset_changes(sender, instance, **kwargs):
+    old_instance = Asset.objects.filter(pk=instance.pk).values().first()
+    if (
+        old_instance
+        or instance.asset_detail_status == "UPDATED"
+        or instance.asset_detail_status == "ASSIGNED"
+        or instance.asset_detail_status == "UNASSIGNED"
+    ):
+        changes = {
+            field: getattr(instance, field)
+            for field in old_instance
+            if field != "asset_uuid"
+        }
+        asset_log_data = json.dumps(changes, indent=4, sort_keys=True, default=str)
+
+        if changes:
+            with transaction.atomic():
+                asset_instance = Asset.objects.select_for_update().get(pk=instance.pk)
+                asset_log_entry = AssetLog.objects.create(
+                    asset_uuid=asset_instance,
+                    asset_log=asset_log_data,
+                )
+                asset_log_entry.save()
